+++ packages/lib/src/components/paywall/Subscriptions.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { twMerge } from 'tailwind-merge';
   import SelectionGroup from './SelectionGroup.svelte';
-  import type { TranslationFunction } from 'src/i18n';
-  import type { PaywallSubscription } from 'src/types/Paywall';
+  import type { TranslationFunction } from '../../i18n';
+  import type { PaywallSubscription } from '../../types/Paywall';
   import Column from '../Column.svelte';
   import Features from '../Features.svelte';
   import Button from '../../Button.wc.svelte';
@@ -142,14 +142,18 @@
           </Column>
 
           {#if url}
-            <Button href={url} class="w-full mt-4" variant="primary">
+            <Button
+              href={url}
+              class="w-full mt-4 bg-[var(--s-paywall-btn-bg-color)] text-[var(--s-paywall-btn-text-color)]"
+              variant="primary"
+            >
               {buttonText || t('continue')}
             </Button>
           {:else}
             {#await getCheckoutUrl(subscription) then checkoutUrl}
               <Button
                 href={checkoutUrl.replace('poId', 'option')}
-                class="w-full mt-4"
+                class="w-full mt-4 bg-[var(--s-paywall-btn-bg-color)] text-[var(--s-paywall-btn-text-color)]"
                 variant="primary"
               >
                 {buttonText || t('continue')}
@@ -160,7 +164,7 @@
       </Column>
     {:else}
       {#if i}
-        <hr class={twMerge('w-full dark:border-gray-800')} />
+        <hr class={twMerge('w-full border-primary opacity-25')} />
       {/if}
       <Selection
         {id}