+++ packages/lib/src/types.ts
@@ -1,4 +1,3 @@
-import type { SesamyAPI } from '@sesamy/sesamy-js';
 import type { HTMLButtonAttributes } from 'svelte/elements';
 
 // NOTE: (important!)