+++ vite.lib.config.ts
@@ -13,10 +13,11 @@ export default defineConfig({
     },
     outDir: 'dist/lib',
     rollupOptions: {
-      external: ['svelte', '@sesamy/sesamy-js'],
+      external: ['@sesamy/sesamy-js', '@sesamy/sdk'],
       output: {
         globals: {
-          svelte: 'Svelte'
+          '@sesamy/sesamy-js': 'SesamyJS',
+          '@sesamy/sdk': 'SesamySDK'
         }
       }
     }